requirements.txt
google.generativeai
llamaapi
PyPDF2
streamlit
